api/requirements.txt
azure-functions
requests
python-dotenv
azure-cognitiveservices-vision-computervision
azure-search-documents
azure-core